Split up the raw and transaction modules
