+++ src/raw/client.rs
@@ -1,0 +1,279 @@
+// Copyright 2019 TiKV Project Authors. Licensed under Apache-2.0.
+
+use super::requests::{
+    BatchDeleteInner, BatchGetInner, BatchPutInner, BatchScanInner, DeleteInner, DeleteRangeInner,
+    GetInner, PutInner, ScanInner,
+};
+use super::{BatchDelete, BatchGet, BatchPut, BatchScan, Delete, DeleteRange, Get, Put, Scan};
+use crate::{rpc::RpcClient, Config, Key, BoundRange, KvPair, Result, Value};
+
+use futures::prelude::*;
+use futures::task::{Context, Poll};
+use std::{pin::Pin, sync::Arc, u32};
+
+/// The TiKV raw [`Client`](Client) is used to issue requests to the TiKV server and PD cluster.
+pub struct Client {
+    rpc: Arc<RpcClient>,
+}
+
+impl Client {
+    /// Create a new [`Client`](Client) once the [`Connect`](Connect) resolves.
+    ///
+    /// ```rust,no_run
+    /// # #![feature(async_await)]
+    /// # use tikv_client::{Config, raw::Client};
+    /// # use futures::prelude::*;
+    /// # futures::executor::block_on(async {
+    /// let connect = Client::connect(Config::default());
+    /// let client = connect.await.unwrap();
+    /// # });
+    /// ```
+    pub fn connect(config: Config) -> Connect {
+        Connect::new(config)
+    }
+
+    #[inline]
+    fn rpc(&self) -> Arc<RpcClient> {
+        Arc::clone(&self.rpc)
+    }
+
+    /// Create a new [`Get`](Get) request.
+    ///
+    /// Once resolved this request will result in the fetching of the value associated with the
+    /// given key.
+    ///
+    /// ```rust,no_run
+    /// # #![feature(async_await)]
+    /// # use tikv_client::{Value, Config, raw::Client};
+    /// # use futures::prelude::*;
+    /// # futures::executor::block_on(async {
+    /// # let connecting_client = Client::connect(Config::new(vec!["192.168.0.100", "192.168.0.101"]));
+    /// # let connected_client = connecting_client.await.unwrap();
+    /// let key = "TiKV";
+    /// let req = connected_client.get(key);
+    /// let result: Option<Value> = req.await.unwrap();
+    /// # });
+    /// ```
+    pub fn get(&self, key: impl Into<Key>) -> Get {
+        Get::new(self.rpc(), GetInner::new(key.into()))
+    }
+
+    /// Create a new [`BatchGet`](BatchGet) request.
+    ///
+    /// Once resolved this request will result in the fetching of the values associated with the
+    /// given keys.
+    ///
+    /// ```rust,no_run
+    /// # #![feature(async_await)]
+    /// # use tikv_client::{KvPair, Config, raw::Client};
+    /// # use futures::prelude::*;
+    /// # futures::executor::block_on(async {
+    /// # let connecting_client = Client::connect(Config::new(vec!["192.168.0.100", "192.168.0.101"]));
+    /// # let connected_client = connecting_client.await.unwrap();
+    /// let keys = vec!["TiKV", "TiDB"];
+    /// let req = connected_client.batch_get(keys);
+    /// let result: Vec<KvPair> = req.await.unwrap();
+    /// # });
+    /// ```
+    pub fn batch_get(&self, keys: impl IntoIterator<Item = impl Into<Key>>) -> BatchGet {
+        BatchGet::new(
+            self.rpc(),
+            BatchGetInner::new(keys.into_iter().map(Into::into).collect()),
+        )
+    }
+
+    /// Create a new [`Put`](Put) request.
+    ///
+    /// Once resolved this request will result in the setting of the value associated with the given key.
+    ///
+    /// ```rust,no_run
+    /// # #![feature(async_await)]
+    /// # use tikv_client::{Key, Value, Config, raw::Client};
+    /// # use futures::prelude::*;
+    /// # futures::executor::block_on(async {
+    /// # let connecting_client = Client::connect(Config::new(vec!["192.168.0.100", "192.168.0.101"]));
+    /// # let connected_client = connecting_client.await.unwrap();
+    /// let key = "TiKV";
+    /// let val = "TiKV";
+    /// let req = connected_client.put(key, val);
+    /// let result: () = req.await.unwrap();
+    /// # });
+    /// ```
+    pub fn put(&self, key: impl Into<Key>, value: impl Into<Value>) -> Put {
+        Put::new(self.rpc(), PutInner::new(key.into(), value.into()))
+    }
+
+    /// Create a new [`BatchPut`](BatchPut) request.
+    ///
+    /// Once resolved this request will result in the setting of the value associated with the given key.
+    ///
+    /// ```rust,no_run
+    /// # #![feature(async_await)]
+    /// # use tikv_client::{Error, Result, KvPair, Key, Value, Config, raw::Client};
+    /// # use futures::prelude::*;
+    /// # futures::executor::block_on(async {
+    /// # let connecting_client = Client::connect(Config::new(vec!["192.168.0.100", "192.168.0.101"]));
+    /// # let connected_client = connecting_client.await.unwrap();
+    /// let kvpair1 = ("PD", "Go");
+    /// let kvpair2 = ("TiKV", "Rust");
+    /// let iterable = vec![kvpair1, kvpair2];
+    /// let req = connected_client.batch_put(iterable);
+    /// let result: () = req.await.unwrap();
+    /// # });
+    /// ```
+    pub fn batch_put(&self, pairs: impl IntoIterator<Item = impl Into<KvPair>>) -> BatchPut {
+        BatchPut::new(
+            self.rpc(),
+            BatchPutInner::new(pairs.into_iter().map(Into::into).collect()),
+        )
+    }
+
+    /// Create a new [`Delete`](Delete) request.
+    ///
+    /// Once resolved this request will result in the deletion of the given key.
+    ///
+    /// ```rust,no_run
+    /// # #![feature(async_await)]
+    /// # use tikv_client::{Key, Config, raw::Client};
+    /// # use futures::prelude::*;
+    /// # futures::executor::block_on(async {
+    /// # let connecting_client = Client::connect(Config::new(vec!["192.168.0.100", "192.168.0.101"]));
+    /// # let connected_client = connecting_client.await.unwrap();
+    /// let key = "TiKV";
+    /// let req = connected_client.delete(key);
+    /// let result: () = req.await.unwrap();
+    /// # });
+    /// ```
+    pub fn delete(&self, key: impl Into<Key>) -> Delete {
+        Delete::new(self.rpc(), DeleteInner::new(key.into()))
+    }
+
+    /// Create a new [`BatchDelete`](BatchDelete) request.
+    ///
+    /// Once resolved this request will result in the deletion of the given keys.
+    ///
+    /// ```rust,no_run
+    /// # #![feature(async_await)]
+    /// # use tikv_client::{Config, raw::Client};
+    /// # use futures::prelude::*;
+    /// # futures::executor::block_on(async {
+    /// # let connecting_client = Client::connect(Config::new(vec!["192.168.0.100", "192.168.0.101"]));
+    /// # let connected_client = connecting_client.await.unwrap();
+    /// let keys = vec!["TiKV", "TiDB"];
+    /// let req = connected_client.batch_delete(keys);
+    /// let result: () = req.await.unwrap();
+    /// # });
+    /// ```
+    pub fn batch_delete(&self, keys: impl IntoIterator<Item = impl Into<Key>>) -> BatchDelete {
+        BatchDelete::new(
+            self.rpc(),
+            BatchDeleteInner::new(keys.into_iter().map(Into::into).collect()),
+        )
+    }
+
+    /// Create a new [`Scan`](Scan) request.
+    ///
+    /// Once resolved this request will result in a scanner over the given keys.
+    ///
+    /// ```rust,no_run
+    /// # #![feature(async_await)]
+    /// # use tikv_client::{KvPair, Config, raw::Client};
+    /// # use futures::prelude::*;
+    /// # futures::executor::block_on(async {
+    /// # let connecting_client = Client::connect(Config::new(vec!["192.168.0.100", "192.168.0.101"]));
+    /// # let connected_client = connecting_client.await.unwrap();
+    /// let inclusive_range = "TiKV"..="TiDB";
+    /// let req = connected_client.scan(inclusive_range, 2);
+    /// let result: Vec<KvPair> = req.await.unwrap();
+    /// # });
+    /// ```
+    pub fn scan(&self, range: impl Into<BoundRange>, limit: u32) -> Scan {
+        Scan::new(self.rpc(), ScanInner::new(range.into(), limit))
+    }
+
+    /// Create a new [`BatchScan`](BatchScan) request.
+    ///
+    /// Once resolved this request will result in a set of scanners over the given keys.
+    ///
+    /// ```rust,no_run
+    /// # #![feature(async_await)]
+    /// # use tikv_client::{Key, Config, raw::Client};
+    /// # use futures::prelude::*;
+    /// # futures::executor::block_on(async {
+    /// # let connecting_client = Client::connect(Config::new(vec!["192.168.0.100", "192.168.0.101"]));
+    /// # let connected_client = connecting_client.await.unwrap();
+    /// let inclusive_range1 = "TiDB"..="TiKV";
+    /// let inclusive_range2 = "TiKV"..="TiSpark";
+    /// let iterable = vec![inclusive_range1, inclusive_range2];
+    /// let req = connected_client.batch_scan(iterable, 2);
+    /// let result = req.await;
+    /// # });
+    /// ```
+    pub fn batch_scan(
+        &self,
+        ranges: impl IntoIterator<Item = impl Into<BoundRange>>,
+        each_limit: u32,
+    ) -> BatchScan {
+        BatchScan::new(
+            self.rpc(),
+            BatchScanInner::new(
+                ranges.into_iter().map(KeyRange::into).collect(),
+                each_limit,
+            ),
+        )
+    }
+
+    /// Create a new [`DeleteRange`](DeleteRange) request.
+    ///
+    /// Once resolved this request will result in the deletion of all keys over the given range.
+    ///
+    /// ```rust,no_run
+    /// # #![feature(async_await)]
+    /// # use tikv_client::{Key, Config, raw::Client};
+    /// # use futures::prelude::*;
+    /// # futures::executor::block_on(async {
+    /// # let connecting_client = Client::connect(Config::new(vec!["192.168.0.100", "192.168.0.101"]));
+    /// # let connected_client = connecting_client.await.unwrap();
+    /// let inclusive_range = "TiKV"..="TiDB";
+    /// let req = connected_client.delete_range(inclusive_range);
+    /// let result: () = req.await.unwrap();
+    /// # });
+    /// ```
+    pub fn delete_range(&self, range: impl Into<BoundRange>) -> DeleteRange {
+        DeleteRange::new(self.rpc(), DeleteRangeInner::new(range.into()))
+    }
+}
+
+/// An unresolved [`Client`](Client) connection to a TiKV cluster.
+///
+/// Once resolved it will result in a connected [`Client`](Client).
+///
+/// ```rust,no_run
+/// # #![feature(async_await)]
+/// use tikv_client::{Config, raw::{Client, Connect}};
+/// use futures::prelude::*;
+///
+/// # futures::executor::block_on(async {
+/// let connect: Connect = Client::connect(Config::default());
+/// let client: Client = connect.await.unwrap();
+/// # });
+/// ```
+pub struct Connect {
+    config: Config,
+}
+
+impl Connect {
+    fn new(config: Config) -> Self {
+        Connect { config }
+    }
+}
+
+impl Future for Connect {
+    type Output = Result<Client>;
+
+    fn poll(self: Pin<&mut Self>, _cx: &mut Context) -> Poll<Self::Output> {
+        let config = &self.config;
+        let rpc = Arc::new(RpcClient::connect(config)?);
+        Poll::Ready(Ok(Client { rpc }))
+    }
+}
